feat: expandable block
